Data-Modeling-Projects/Distributed NoSQL/requirements.txt
cassandra-driver>=3.25.0
pandas>=1.3.0
jupyter>=1.0.0
ipykernel>=6.0.0
